chore: Refactor atomics support
- Remove `--disable-c11' scons option. C11 and C++11 are
  used unconditionally, so this flag is not needed.

- Earlier, `--disable-c11' enabled `libatomic_ops'.
  Now we enable `libatomic_ops' on non-GNU targets.

  On GNU-like targets (gcc, clang, etc), we use GCC builtin
  atomics (__atomic_xxx).

  We don't use C11 or C++11 atomics (std::atomic_xxx and _Atomic)
  because they don't allow atomic operations on raw memory.

- Update `libatomic_ops' version of `AtomicOps' to be on par
  with the GNU one. Also simplify & cleanup generated code.

- Replace Atomic<T> with a few classes: AtomicInt<T>, AtomicBool
  AtomicSize, AtomicPtr<T>.

  Each has different set of methods.

  AtomicInt doesn't allow 64-bit atomics to discourage writing
  code that would work differently on 32-bit platforms.

  For portable pointer-size-dependent atomic, use AtomicSize.
  For portable atomic boolean, use AtomicBool.
  For portable atomic pointer, use AtomicPtr.
  For portable 64-bit atomic-like type, Seqlock can be used.

  For non-portable code, AtomicOps can be used directly
  instead of AtomicInt wrapper.

- Expand tests for atomics.

- Add null pointer panic to AtomicPtr.

+++ src/internal_modules/roc_core/heap_arena.cpp
@@ -115,7 +115,7 @@ void HeapArena::deallocate(void* ptr) {
         }
     }
 
-    const int n = num_allocations_--;
+    const size_t n = num_allocations_--;
     if (n == 0) {
         roc_panic("heap arena: unpaired deallocate");
     }
